addressed linting, correct config
