+++ package.json
@@ -22,6 +22,8 @@
   },
   "dependencies": {
     "@babel/eslint-parser": "^7.16.5",
+    "aws-sdk": "^2.1048.0",
+    "axios": "^0.24.0",
     "babel": "^6.23.0",
     "babel-eslint": "^10.1.0"
   }
